create useCryptoHandler to offload fetching basic crypto data for table
